requirement.txt
numpy
matplotlib
torch
torchvision
transformers
pathlib
pillow 
flask>=3.0
torch
torchvision
pillow
